+++ packages/zendesk-client/src/use-managed-zendesk-chat.tsx
@@ -20,75 +20,27 @@ import SmoochLibrary from 'smooch';
 import { AttachmentMessage } from './components/attachment-message';
 import { CSATForm } from './components/csat-form';
 import { SMOOCH_INTEGRATION_ID, SMOOCH_INTEGRATION_ID_STAGING } from './constants';
-import { ZendeskConversation } from './types';
+import {
+	ConversationData,
+	ZendeskConversation,
+	ZendeskImagePreview,
+	ZendeskUploadingImage,
+} from './types';
 import { useAttachFileToConversation } from './use-attach-file';
 import {
 	useAuthenticateZendeskMessaging,
 	fetchMessagingAuth,
 } from './use-authenticate-zendesk-messaging';
-import { isTestModeEnvironment, convertZendeskMessageToAgentticFormat } from './util';
+import {
+	convertZendeskMessageToAgentticFormat,
+	getSmoochContainer,
+	isSupportedImageType,
+	isTestModeEnvironment,
+	MAX_ATTACHMENTS,
+	playNotificationSound,
+	SUPPORTED_IMAGE_TYPES,
+} from './util';
 import type { AgentticMessage, ZendeskMessage } from './types';
-
-const SUPPORTED_IMAGE_TYPES = [ 'image/jpeg', 'image/jpg', 'image/png', 'image/gif' ];
-const MAX_ATTACHMENTS = 5;
-
-function isSupportedImageType( type: string ) {
-	return SUPPORTED_IMAGE_TYPES.includes( type );
-}
-
-/** Minimal image preview shape for attachment upload UI (compatible with UseImageUploadResult). */
-export type ZendeskImagePreview = {
-	id: string;
-	url: string;
-	name: string;
-	alt: string;
-	mime_type: string;
-	file: File;
-};
-
-/** Minimal uploading image shape (compatible with UseImageUploadResult.uploadingImages). */
-export type ZendeskUploadingImage = {
-	id: string;
-	url?: string;
-	name?: string;
-};
-
-type ConversationData = {
-	conversation: {
-		id: string;
-	};
-};
-
-let smoochContainer: HTMLDivElement | null = null;
-
-function getSmoochContainer(): HTMLDivElement | null {
-	if ( typeof document === 'undefined' ) {
-		return null;
-	}
-
-	const existing = document.querySelector< HTMLDivElement >( '.smooch-container' );
-	if ( existing ) {
-		smoochContainer = existing;
-	} else if ( ! smoochContainer ) {
-		smoochContainer = document.createElement( 'div' );
-		smoochContainer.className = 'smooch-container';
-	}
-
-	// Keep the container hidden since we're using embedded mode.
-	smoochContainer.style.display = 'none';
-	smoochContainer.style.position = 'absolute';
-	smoochContainer.style.top = '0';
-	smoochContainer.style.left = '0';
-	smoochContainer.style.width = '100%';
-	smoochContainer.style.height = '100%';
-	smoochContainer.style.zIndex = '1000';
-
-	if ( ! document.body.contains( smoochContainer ) ) {
-		document.body.appendChild( smoochContainer );
-	}
-
-	return smoochContainer;
-}
 
 function useSmooch( enabled = true ) {
 	const queryClient = useQueryClient();
@@ -141,29 +93,6 @@ function useSmooch( enabled = true ) {
 
 	return { ...smoochQuery, isLoading: isAuthenticatingZendeskMessaging || smoochQuery.isFetching };
 }
-
-const playNotificationSound = () => {
-	// @ts-expect-error expected because of fallback webkitAudioContext
-	const audioContext = new ( window.AudioContext || window.webkitAudioContext )();
-
-	const duration = 0.7;
-	const oscillator = audioContext.createOscillator();
-	const gainNode = audioContext.createGain();
-
-	// Configure oscillator
-	oscillator.type = 'sine';
-	oscillator.frequency.setValueAtTime( 660, audioContext.currentTime );
-
-	// Configure gain for a smoother fade-out
-	gainNode.gain.setValueAtTime( 0.3, audioContext.currentTime );
-	gainNode.gain.exponentialRampToValueAtTime( 0.001, audioContext.currentTime + duration );
-
-	// Connect & start
-	oscillator.connect( gainNode );
-	gainNode.connect( audioContext.destination );
-	oscillator.start();
-	oscillator.stop( audioContext.currentTime + duration );
-};
 /**
  * Returns a complete API for managing a Zendesk chat.
  * @returns An object with the following properties:
